feat: modularize App.jsx, replace IndexedDB with Dexie.js, add Vite code splitting
