+++ package.json
@@ -19,6 +19,7 @@
     "@tauri-apps/plugin-opener": "^2.5.3",
     "@vercel/analytics": "^1.6.1",
     "clsx": "^2.1.1",
+    "dexie": "^4.3.0",
     "framer-motion": "^12.23.26",
     "html2canvas": "^1.4.1",
     "lucide-react": "^0.300.0",
