+++ src/components/modals/index.js
@@ -10,4 +10,6 @@ export { CategoryManagerModal } from './CategoryManagerModal';
 export { ConfirmModal } from './ConfirmModal';
 export { AddTemplateTypeModal } from './AddTemplateTypeModal';
 export { VideoSubTypeModal } from './VideoSubTypeModal';
-export { LinkTemplateModal } from './LinkTemplateModal';
+export { LinkTemplateModal } from './LinkTemplateModal';export { NoticeModal } from './NoticeModal';
+export { ExportTemplatesModal } from './ExportTemplatesModal';
+export { ImageUrlModal } from './ImageUrlModal';
